Rename package login to hello, rename mvp names in new hello packag
