+++ app/src/main/java/thereisnospon/acclient/DebugActivity.java
@@ -16,9 +16,9 @@ import java.util.List;
 import thereisnospon.acclient.base.activity.DrawerActivity;
 import thereisnospon.acclient.event.Arg;
 import thereisnospon.acclient.event.Msg;
-import thereisnospon.acclient.modules.login.LoginActivity;
+import thereisnospon.acclient.modules.hello.LoginActivity;
 
-import thereisnospon.acclient.modules.login.LoginUtil;
+import thereisnospon.acclient.modules.hello.LoginUtil;
 import thereisnospon.acclient.modules.problem_list.HdojActivity;
 import thereisnospon.acclient.modules.rank.RankActivity;
 import thereisnospon.acclient.modules.search_people.SearchPeopleActivity;
