+++ app/src/main/res/layout/activity_hello_scene_index.xml
@@ -13,7 +13,7 @@
         android:alpha="1"
         android:background="@color/colorPrimary"
         android:orientation="vertical"
-        tools:context="thereisnospon.acclient.modules.login.LoginActivity">
+        tools:context="thereisnospon.acclient.modules.hello.LoginActivity">
 
         <ImageView
             style="@style/HelloLogo"/>
