+++ app/src/main/res/layout/activity_hello_scene_login.xml
@@ -16,7 +16,7 @@
             android:layout_height="wrap_content"
             android:alpha="1"
             android:orientation="vertical"
-            tools:context="thereisnospon.acclient.modules.login.LoginActivity">
+            tools:context="thereisnospon.acclient.modules.hello.LoginActivity">
 
             <ImageView
                 style="@style/HelloLogo"/>
